+++ src/propulsionai/resources/models.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from typing import Any, Dict, Callable, Iterable, Optional
+from typing import List, Iterable, Optional
 
 import httpx
 
@@ -45,8 +45,10 @@ class ModelsResource(SyncAPIResource):
         model: str,
         stream: bool,
         wait: bool | NotGiven = NOT_GIVEN,
+        knowledgebases: List[str] | NotGiven = NOT_GIVEN,
         max_tokens: Optional[int] | NotGiven = NOT_GIVEN,
         n: Optional[int] | NotGiven = NOT_GIVEN,
+        task_id: str | NotGiven = NOT_GIVEN,
         temperature: Optional[float] | NotGiven = NOT_GIVEN,
         tool_choice: model_chat_params.ToolChoice | NotGiven = NOT_GIVEN,
         tools: Iterable[model_chat_params.Tool] | NotGiven = NOT_GIVEN,
@@ -64,9 +66,13 @@ class ModelsResource(SyncAPIResource):
         Args:
           wait: Whether to wait for the response or not.
 
+          knowledgebases: A list of knowledgebase IDs to use in the model.
+
           max_tokens: The maximum number of tokens that can be generated in the chat completion.
 
           n: How many chat completion choices to generate for each input message.
+
+          task_id: Optional task ID associated with the request.
 
           temperature: An alternative to sampling with temperature, called nucleus sampling.
 
@@ -98,8 +104,10 @@ class ModelsResource(SyncAPIResource):
                     "messages": messages,
                     "model": model,
                     "stream": stream,
+                    "knowledgebases": knowledgebases,
                     "max_tokens": max_tokens,
                     "n": n,
+                    "task_id": task_id,
                     "temperature": temperature,
                     "tool_choice": tool_choice,
                     "tools": tools,
@@ -125,8 +133,10 @@ class ModelsResource(SyncAPIResource):
         model: str,
         stream: bool,
         wait: bool | NotGiven = NOT_GIVEN,
+        knowledgebases: List[str] | NotGiven = NOT_GIVEN,
         max_tokens: Optional[int] | NotGiven = NOT_GIVEN,
         n: Optional[int] | NotGiven = NOT_GIVEN,
+        task_id: str | NotGiven = NOT_GIVEN,
         temperature: Optional[float] | NotGiven = NOT_GIVEN,
         tool_choice: model_chat_params.ToolChoice | NotGiven = NOT_GIVEN,
         tools: Iterable[model_chat_params.Tool] | NotGiven = NOT_GIVEN,
@@ -137,48 +147,55 @@ class ModelsResource(SyncAPIResource):
         extra_query: Query | None = None,
         extra_body: Body | None = None,
         timeout: float | httpx.Timeout | None | NotGiven = NOT_GIVEN,
-        available_function_map: Dict[str, Callable[..., Any]],
-    ) -> ModelChatResponse:
-        if not model_id:
-            raise ValueError(f"Expected a non-empty value for `model_id` but received {model_id!r}")
-        
-        if not tools or not available_function_map:
-            initial_response: ModelChatResponse = self._post(
-                f"/api/v1/{model_id}/run",
-                body=maybe_transform(
-                    {
-                        "messages": messages,
-                        "model": model,
-                        "stream": stream,
-                        "max_tokens": max_tokens,
-                        "n": n,
-                        "temperature": temperature,
-                        "tool_choice": tool_choice,
-                        "tools": tools,
-                        "top_p": top_p,
-                    },
-                    model_chat_params.ModelChatParams,
-                ),
-                options=make_request_options(
-                    extra_headers=extra_headers,
-                    extra_query=extra_query,
-                    extra_body=extra_body,
-                    timeout=timeout,
-                    query=maybe_transform({"wait": wait}, model_chat_params.ModelChatParams),
-                ),
-                cast_to=ModelChatResponse,
-            )
-            return initial_response
-        
-        initial_response = self._post(
-            f"/api/v1/{model_id}/run",
+    ) -> ModelEpResponse:
+        """
+        Call a deployment endpoint with specified tools and messages.
+
+        Args:
+          wait: Whether to wait for the response or not.
+
+          knowledgebases: A list of knowledgebase IDs to use in the model.
+
+          max_tokens: The maximum number of tokens that can be generated in the chat completion.
+
+          n: How many chat completion choices to generate for each input message.
+
+          task_id: Optional task ID associated with the request.
+
+          temperature: An alternative to sampling with temperature, called nucleus sampling.
+
+          tool_choice: Controls which (if any) tool is called by the model. `none` means the model will
+              not call any tool and instead generates a message. `auto` means the model can
+              pick between generating a message or calling one or more tools. `required` means
+              the model must call one or more tools.
+
+          tools: A list of tools the model may call. Currently, only functions are supported as a
+              tool. Use this to provide a list of functions the model may generate JSON inputs
+              for. A max of 128 functions are supported.
+
+          top_p: An alternative to sampling with temperature, called nucleus sampling.
+
+          extra_headers: Send extra headers
+
+          extra_query: Add additional query parameters to the request
+
+          extra_body: Add additional JSON properties to the request
+
+          timeout: Override the client-level default timeout for this request, in seconds
+        """
+        if not deployment_tag:
+            raise ValueError(f"Expected a non-empty value for `deployment_tag` but received {deployment_tag!r}")
+        return self._post(
+            f"/api/v1/chat/{deployment_tag}",
             body=maybe_transform(
                 {
                     "messages": messages,
                     "model": model,
                     "stream": stream,
+                    "knowledgebases": knowledgebases,
                     "max_tokens": max_tokens,
                     "n": n,
+                    "task_id": task_id,
                     "temperature": temperature,
                     "tool_choice": tool_choice,
                     "tools": tools,
@@ -269,8 +286,10 @@ class AsyncModelsResource(AsyncAPIResource):
         model: str,
         stream: bool,
         wait: bool | NotGiven = NOT_GIVEN,
+        knowledgebases: List[str] | NotGiven = NOT_GIVEN,
         max_tokens: Optional[int] | NotGiven = NOT_GIVEN,
         n: Optional[int] | NotGiven = NOT_GIVEN,
+        task_id: str | NotGiven = NOT_GIVEN,
         temperature: Optional[float] | NotGiven = NOT_GIVEN,
         tool_choice: model_chat_params.ToolChoice | NotGiven = NOT_GIVEN,
         tools: Iterable[model_chat_params.Tool] | NotGiven = NOT_GIVEN,
@@ -288,9 +307,13 @@ class AsyncModelsResource(AsyncAPIResource):
         Args:
           wait: Whether to wait for the response or not.
 
+          knowledgebases: A list of knowledgebase IDs to use in the model.
+
           max_tokens: The maximum number of tokens that can be generated in the chat completion.
 
           n: How many chat completion choices to generate for each input message.
+
+          task_id: Optional task ID associated with the request.
 
           temperature: An alternative to sampling with temperature, called nucleus sampling.
 
@@ -322,8 +345,10 @@ class AsyncModelsResource(AsyncAPIResource):
                     "messages": messages,
                     "model": model,
                     "stream": stream,
+                    "knowledgebases": knowledgebases,
                     "max_tokens": max_tokens,
                     "n": n,
+                    "task_id": task_id,
                     "temperature": temperature,
                     "tool_choice": tool_choice,
                     "tools": tools,
@@ -339,6 +364,94 @@ class AsyncModelsResource(AsyncAPIResource):
                 query=await async_maybe_transform({"wait": wait}, model_chat_params.ModelChatParams),
             ),
             cast_to=ModelChatResponse,
+        )
+
+    async def ep(
+        self,
+        deployment_tag: str,
+        *,
+        messages: Iterable[model_ep_params.Message],
+        model: str,
+        stream: bool,
+        wait: bool | NotGiven = NOT_GIVEN,
+        knowledgebases: List[str] | NotGiven = NOT_GIVEN,
+        max_tokens: Optional[int] | NotGiven = NOT_GIVEN,
+        n: Optional[int] | NotGiven = NOT_GIVEN,
+        task_id: str | NotGiven = NOT_GIVEN,
+        temperature: Optional[float] | NotGiven = NOT_GIVEN,
+        tool_choice: model_ep_params.ToolChoice | NotGiven = NOT_GIVEN,
+        tools: Iterable[model_ep_params.Tool] | NotGiven = NOT_GIVEN,
+        top_p: Optional[float] | NotGiven = NOT_GIVEN,
+        # Use the following arguments if you need to pass additional parameters to the API that aren't available via kwargs.
+        # The extra values given here take precedence over values defined on the client or passed to this method.
+        extra_headers: Headers | None = None,
+        extra_query: Query | None = None,
+        extra_body: Body | None = None,
+        timeout: float | httpx.Timeout | None | NotGiven = NOT_GIVEN,
+    ) -> ModelEpResponse:
+        """
+        Call a deployment endpoint with specified tools and messages.
+
+        Args:
+          wait: Whether to wait for the response or not.
+
+          knowledgebases: A list of knowledgebase IDs to use in the model.
+
+          max_tokens: The maximum number of tokens that can be generated in the chat completion.
+
+          n: How many chat completion choices to generate for each input message.
+
+          task_id: Optional task ID associated with the request.
+
+          temperature: An alternative to sampling with temperature, called nucleus sampling.
+
+          tool_choice: Controls which (if any) tool is called by the model. `none` means the model will
+              not call any tool and instead generates a message. `auto` means the model can
+              pick between generating a message or calling one or more tools. `required` means
+              the model must call one or more tools.
+
+          tools: A list of tools the model may call. Currently, only functions are supported as a
+              tool. Use this to provide a list of functions the model may generate JSON inputs
+              for. A max of 128 functions are supported.
+
+          top_p: An alternative to sampling with temperature, called nucleus sampling.
+
+          extra_headers: Send extra headers
+
+          extra_query: Add additional query parameters to the request
+
+          extra_body: Add additional JSON properties to the request
+
+          timeout: Override the client-level default timeout for this request, in seconds
+        """
+        if not deployment_tag:
+            raise ValueError(f"Expected a non-empty value for `deployment_tag` but received {deployment_tag!r}")
+        return await self._post(
+            f"/api/v1/chat/{deployment_tag}",
+            body=await async_maybe_transform(
+                {
+                    "messages": messages,
+                    "model": model,
+                    "stream": stream,
+                    "knowledgebases": knowledgebases,
+                    "max_tokens": max_tokens,
+                    "n": n,
+                    "task_id": task_id,
+                    "temperature": temperature,
+                    "tool_choice": tool_choice,
+                    "tools": tools,
+                    "top_p": top_p,
+                },
+                model_ep_params.ModelEpParams,
+            ),
+            options=make_request_options(
+                extra_headers=extra_headers,
+                extra_query=extra_query,
+                extra_body=extra_body,
+                timeout=timeout,
+                query=await async_maybe_transform({"wait": wait}, model_ep_params.ModelEpParams),
+            ),
+            cast_to=ModelEpResponse,
         )