+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## 0.1.0-alpha.7 (2024-07-03)
+## 0.1.0-alpha.7 (2024-07-09)
 
 Full Changelog: [v0.1.0-alpha.6...v0.1.0-alpha.7](https://github.com/propulsion-ai/propulsionai-python/compare/v0.1.0-alpha.6...v0.1.0-alpha.7)
 
@@ -8,6 +8,8 @@ Full Changelog: [v0.1.0-alpha.6...v0.1.0-alpha.7](https://github.com/propulsion-
 
 * **api:** OpenAPI spec update via Stainless API ([#23](https://github.com/propulsion-ai/propulsionai-python/issues/23)) ([828e954](https://github.com/propulsion-ai/propulsionai-python/commit/828e954676813291c760ff43c3b69718661d2af7))
 * **api:** OpenAPI spec update via Stainless API ([#25](https://github.com/propulsion-ai/propulsionai-python/issues/25)) ([7c77e0f](https://github.com/propulsion-ai/propulsionai-python/commit/7c77e0f2d32c04b0eab805ffac25f17b8cd18a7f))
+* **api:** OpenAPI spec update via Stainless API ([#31](https://github.com/propulsion-ai/propulsionai-python/issues/31)) ([c6ee599](https://github.com/propulsion-ai/propulsionai-python/commit/c6ee5998a353045b5c2dab7a429b595e6c90ea1c))
+* **api:** OpenAPI spec update via Stainless API ([#33](https://github.com/propulsion-ai/propulsionai-python/issues/33)) ([56bbd89](https://github.com/propulsion-ai/propulsionai-python/commit/56bbd894fe491f40dcfe9c2f7d6727b5d6651338))
 
 ## 0.1.0-alpha.6 (2024-06-12)